created custom component FilterSelector
