+++ frontend/src/components/Search/FilterSelector.css
@@ -1,0 +1,37 @@
+.FilterSelector {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 170px;
+  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+  background-color: #1a1a1a;
+  border: 0;
+  border-left: 1px solid var(--bgColor);
+  border-right: 1px solid var(--bgColor);
+}
+
+.FilterSelector:hover {
+  border: 1px solid var(--buttonHoverBorderColor);
+}
+
+.FilterSelector select {
+  outline: none;
+  background-color: inherit;
+  border: 0;
+  height: 100%;
+  width: 100%;
+  text-align: center;
+  cursor: pointer;
+  padding: 0 1.3em .1em 0;
+  font-size: .9em;
+
+  /* hide default select icon */
+  appearance: none;
+  -webkit-appearance: none;
+  -moz-appearance: none;
+}
+
+.FilterSelector svg {
+  position: absolute;
+  pointer-events: none;
+}
