+++ frontend/src/components/Search/Search.css
@@ -43,43 +43,6 @@
   border-color: var(--buttonHoverBorderColor);
 }
 
-.Search__FilterSelector {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 170px;
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-  background-color: #1a1a1a;
-  border: 0;
-  border-left: 1px solid var(--bgColor);
-  border-right: 1px solid var(--bgColor);
-}
-
-.Search__FilterSelector:hover {
-  border: 1px solid var(--buttonHoverBorderColor);
-}
-
-.Search__FilterSelector select {
-  outline: none;
-  background-color: inherit;
-  border: 0;
-  height: 100%;
-  width: 100%;
-  text-align: center;
-  cursor: pointer;
-  padding-right: 1.3em;
-
-  /* hide default select icon */
-  appearance: none;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-}
-
-.Search__FilterSelector svg {
-  position: absolute;
-  pointer-events: none;
-}
-
 .Search button {
   border-top-left-radius: 0;
   border-bottom-left-radius: 0;
